[gene.y] 인증 관리 Step1, 2 (#1)
* feat: 토큰 발급하는 API 생성

* feat: 토큰을 이용하여 본인 정보 응답하기

* AuthenticationPrincipal annotation 생성, configuration 적용

* feat: 예약 생성 인증 적용

* feat: 예약 삭제 인증 적용

* chore: fix typo

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,8 +1,10 @@
 package nextstep.reservation;
 
+import nextstep.member.Member;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
-import nextstep.support.DuplicateEntityException;
+import nextstep.support.exception.AuthorizationException;
+import nextstep.support.exception.DuplicateEntityException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
@@ -21,8 +23,8 @@ public class ReservationService {
         this.scheduleDao = scheduleDao;
     }
 
-    public Long create(ReservationRequest reservationRequest) {
-        Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
+    public Long create(Long scheduleId, Member member) {
+        Schedule schedule = scheduleDao.findById(scheduleId);
         if (schedule == null) {
             throw new NullPointerException();
         }
@@ -34,7 +36,7 @@ public class ReservationService {
 
         Reservation newReservation = new Reservation(
                 schedule,
-                reservationRequest.getName()
+                member.getId()
         );
 
         return reservationDao.save(newReservation);
@@ -49,12 +51,14 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, Member member) {
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
             throw new NullPointerException();
         }
-
+        if (!reservation.getMemberId().equals(member.getId())) {
+            throw new AuthorizationException();
+        }
         reservationDao.deleteById(id);
     }
 }
